Made the Sort interface use generics
Added generic types to the Sort interface, and all the particular sorts
which implement it, to ensure that whatever elements are in the
Comparable[] can actually be compared: i.e. the signature is now <T
extends Comparable<T>>, and a T[] is passed into the sort.

Also, added comments for implementing caching of InsertionSort
instances by the class their generics support.